src/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
autotiktokenizer==0.2.2
azure-ai-documentintelligence==1.0.0b4
azure-common==1.1.28
azure-core==1.32.0
azure-functions==1.21.3
azure-functions-durable==1.2.10
azure-identity==1.19.0
azure-search-documents==11.6.0b6
azure-storage-blob==12.24.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
chonkie==0.4.1
cryptography==44.0.0
debugpy==1.8.11
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
furl==2.1.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
isodate==0.7.2
jiter==0.8.2
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
numpy==2.2.1
openai==1.59.7
orderedmultidict==1.0.1
packaging==24.2
portalocker==2.10.1
propcache==0.2.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3